+++ src/main/resources/db/migration/V1.5__create_table_area.sql
@@ -6,7 +6,7 @@ CREATE TABLE `area` (
   `n_city_provincial_id` bigint(11) NOT NULL COMMENT 'id cua ban ghi tinh thanh',
   `n_county_district_id` bigint(11) COMMENT 'id cua ban ghi quan huyen',
   `n_commune_id` bigint(11) COMMENT 'id cua ban ghi phuong xa',
-  `n_status` int(1) NOT NULL DEFAULT '1' COMMENT 'trang thai ban ghi',
+  `n_status` tinyint(1) NOT NULL DEFAULT '1' COMMENT 'trang thai ban ghi',
   `d_created_at` datetime NOT NULL,
   `d_updated_at` datetime,
   PRIMARY KEY (`n_id`),
